pyproject.toml
[project]
name = "bootcamp-modern-data-stack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.5"
dependencies = [
    "apache-airflow>=3.1.1",
]
